`Cockpit NG`: Added field level code contribution for wizards using POJO & Java 14 Record
