+++ src/com/intellij/idea/plugin/hybris/system/cockpitng/codeInsight/completion/provider/CngFlowPropertyQualifierCompletionProvider.kt
@@ -19,14 +19,49 @@ package com.intellij.idea.plugin.hybris.system.cockpitng.codeInsight.completion.
 
 import com.intellij.codeInsight.completion.CompletionParameters
 import com.intellij.codeInsight.completion.CompletionProvider
+import com.intellij.codeInsight.completion.CompletionResultSet
+import com.intellij.codeInsight.completion.JavaLookupElementBuilder
+import com.intellij.idea.plugin.hybris.common.HybrisConstants
+import com.intellij.idea.plugin.hybris.java.psi.JavaPsiHelper
+import com.intellij.idea.plugin.hybris.project.utils.PluginCommon
 import com.intellij.idea.plugin.hybris.system.cockpitng.psi.CngPsiHelper
 import com.intellij.idea.plugin.hybris.system.type.codeInsight.completion.provider.AttributeDeclarationCompletionProvider
 import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.project.Project
+import com.intellij.psi.JavaPsiFacade
 import com.intellij.psi.PsiElement
+import com.intellij.psi.search.GlobalSearchScope
+import com.intellij.util.ProcessingContext
 
 class CngFlowPropertyQualifierCompletionProvider : AttributeDeclarationCompletionProvider() {
 
     override fun resolveType(element: PsiElement) = CngPsiHelper.resolveContextTypeForNewItemInWizardFlow(element)
+
+    override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
+        val project = parameters.editor.project ?: return
+        val type = resolveType(parameters.position) ?: return
+
+        if (type.contains(".")
+            && type != HybrisConstants.COCKPIT_NG_INITIALIZE_CONTEXT_TYPE
+            && PluginCommon.isPluginActive(PluginCommon.JAVA_PLUGIN_ID)
+        ) addJavaPojoCompletions(project, type, result)
+        else super.addCompletions(parameters, context, result)
+    }
+
+    private fun addJavaPojoCompletions(project: Project, className: String, result: CompletionResultSet) {
+        JavaPsiFacade.getInstance(project)
+            .findClass(className, GlobalSearchScope.allScope(project))
+            ?.let { psiClass ->
+                val fields = psiClass.fields
+
+                return@let if (psiClass.isRecord) fields.toList()
+                else fields
+                    .filter { JavaPsiHelper.hasGetter(psiClass, it) && JavaPsiHelper.hasSetter(psiClass, it) }
+            }
+            ?.map { JavaLookupElementBuilder.forField(it) }
+            ?.forEach { result.addElement(it) }
+    }
+
 
     companion object {
         val instance: CompletionProvider<CompletionParameters> =
